Fix db-core map-ownership corruption from long m_alMaps on LP64
CPeer stored the per-peer map-ownership list as `long m_alMaps[32]`, but
the wire structs it is copied to/from are int32_t[32] (TPacketGDSetup.alMaps
and TMapLocation.alMaps, 128 bytes). On x86 (ILP32) long==int32 so this was
latent; on LP64 (FreeBSD amd64) long is 8 bytes, making m_alMaps 256 bytes
and breaking both copies:

  - CPeer::SetMaps(int32_t*) does memcpy(m_alMaps, pl, sizeof(m_alMaps)):
    256 bytes copied from the 128-byte TPacketGDSetup.alMaps source -> a
    128-byte over-read past the inbound DG_SETUP packet.
  - ClientManager map-location replies memcpy 128 bytes from the long[32]
    (8-byte elements) into TMapLocation.alMaps (int32_t[32]): each small map
    index N stored as a long becomes [N,0] in the int32 view, so the game
    consumer loop (input_db.cpp: `if (0==alMaps[i]) break;`) stops after the
    first map and the rest of the channel's map list is lost.

Store the list as int32_t[32] and return int32_t* from GetMaps() so the
internal storage matches the wire structs exactly on every data model.
No other code assumes long width for GetMaps()/m_alMaps (the 4 memcpy
sites size off the int32_t destination; map_allow_copy already takes
int32_t*). db core rebuilt on FreeBSD amd64.

+++ Lead-Server-Source/db/src/Peer.h
@@ -53,7 +53,7 @@ class CPeer : public CPeerBase
 	WORD	GetP2PPort() { return m_wP2PPort; }
 
 	void	SetMaps(int32_t* pl);
-	long *	GetMaps() { return &m_alMaps[0]; }
+	int32_t *	GetMaps() { return &m_alMaps[0]; }
 
 	bool	SetItemIDRange(TItemIDRangeTable itemRange);
 	bool	SetSpareItemIDRange(TItemIDRangeTable itemRange);
@@ -68,7 +68,7 @@ class CPeer : public CPeerBase
 	DWORD	m_dwUserCount;
 	WORD	m_wListenPort;	// Game server for clients listen port
 	WORD	m_wP2PPort;	// Game server is game server P2P for connection listen port
-	long	m_alMaps[32];	// What maps are you in charge of? ?
+	int32_t	m_alMaps[32];	// What maps are you in charge of? (must match the int32_t[32] wire structs TPacketGDSetup/TMapLocation; bare long was 8 bytes on LP64 -> over-read on SetMaps + mis-packed copy to TMapLocation)
 
 	TItemIDRangeTable m_itemRange;
 	TItemIDRangeTable m_itemSpareRange;
